# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aios Practice Platform API
  version: 1.0.0
  description: 题练通一期正式 OpenAPI 基线。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /schools:
    get:
      tags: [Org]
      summary: 学校列表
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: keyword
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Org]
      summary: 创建学校
      requestBody:
        $ref: '#/components/requestBodies/SchoolBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /schools/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Org]
      summary: 学校详情
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    put:
      tags: [Org]
      summary: 更新学校
      requestBody:
        $ref: '#/components/requestBodies/SchoolBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /schools/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Org]
      summary: 禁用学校
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /grades:
    get:
      tags: [Org]
      summary: 年级列表
      parameters:
        - in: query
          name: school_id
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Org]
      summary: 创建年级
      requestBody:
        $ref: '#/components/requestBodies/GradeBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /grades/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Org]
      summary: 年级详情
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    put:
      tags: [Org]
      summary: 更新年级
      requestBody:
        $ref: '#/components/requestBodies/GradeBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /grades/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Org]
      summary: 禁用年级
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /classes:
    get:
      tags: [Org]
      summary: 班级列表
      parameters:
        - in: query
          name: school_id
          schema:
            type: integer
            format: int64
        - in: query
          name: grade_id
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Org]
      summary: 创建班级
      requestBody:
        $ref: '#/components/requestBodies/ClassBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /classes/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Org]
      summary: 班级详情
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    put:
      tags: [Org]
      summary: 更新班级
      requestBody:
        $ref: '#/components/requestBodies/ClassBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /classes/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Org]
      summary: 禁用班级
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /courses:
    get:
      tags: [Org]
      summary: 课程列表
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: active_at
          schema: { type: string, format: date-time }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Org]
      summary: 创建课程
      requestBody:
        $ref: '#/components/requestBodies/CourseBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /courses/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Org]
      summary: 课程详情
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    put:
      tags: [Org]
      summary: 更新课程
      requestBody:
        $ref: '#/components/requestBodies/CourseBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /courses/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Org]
      summary: 禁用课程
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
...
  requestBodies:
...
    SchoolBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SchoolInput'
    GradeBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GradeInput'
    ClassBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClassInput'
    CourseBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CourseInput'
...
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        request_id:
          type: string
...
    SchoolInput:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
    GradeInput:
      type: object
      required: [school_id, code, name, grade_level]
      properties:
        school_id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        grade_level:
          type: integer
        school_year:
          type: string
          nullable: true
    ClassInput:
      type: object
      required: [school_id, grade_id, code, name]
      properties:
        school_id:
          type: integer
          format: int64
        grade_id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        class_no:
          type: integer
          nullable: true
...
    CourseInput:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        start_at:
          type: string
          format: date-time
          nullable: true
        end_at:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
...